// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  ADMIN: 'ADMIN',
  GURU: 'GURU'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const RegistrationStatus = {
  MENUNGGU: 'MENUNGGU',
  DITERIMA: 'DITERIMA',
  DITOLAK: 'DITOLAK'
} as const

export type RegistrationStatus = (typeof RegistrationStatus)[keyof typeof RegistrationStatus]


export const FieldType = {
  TEXT: 'TEXT',
  TEXTAREA: 'TEXTAREA',
  NUMBER: 'NUMBER',
  FILE: 'FILE',
  SELECT: 'SELECT',
  CHECKBOX: 'CHECKBOX',
  RADIO: 'RADIO'
} as const

export type FieldType = (typeof FieldType)[keyof typeof FieldType]


export const SubmissionStatus = {
  MENUNGGU: 'MENUNGGU',
  DITERIMA: 'DITERIMA',
  DITOLAK: 'DITOLAK'
} as const

export type SubmissionStatus = (typeof SubmissionStatus)[keyof typeof SubmissionStatus]


export const AchievementStatus = {
  MENUNGGU: 'MENUNGGU',
  TERVERIFIKASI: 'TERVERIFIKASI',
  DITOLAK: 'DITOLAK'
} as const

export type AchievementStatus = (typeof AchievementStatus)[keyof typeof AchievementStatus]